LayoutTests/imported/w3c/web-platform-tests/web-animations/responsive/offsetRotate-expected.txt


FAIL offsetRotate responsive to inherited offsetRotate changes assert_equals: expected "auto 200deg" but got "auto 150deg"
